requirements.txt
python-nmap
mcstatus